pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dev-activity-logger"
version = "1.0.0"
description = "Automatically track and analyse your development activity."
requires-python = ">=3.10"
dependencies = [
    "sqlalchemy>=2.0",
    "typer[all]>=0.9",
    "rich>=13.0",
    "pyyaml>=6.0",
    "gitpython>=3.1",
]

[project.scripts]
devlog = "cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["cli*", "collector*", "config*", "daemon*", "db*", "processor*"]

